Moved many class instances previously located directly in EngineMain into an internal struct inside there

+++ Oxygen/soncthrickles/source/sonic3air/menu/GameMenuBase.cpp
@@ -366,7 +366,7 @@ bool GameMenuEntries::sanitizeSelectedIndex(bool allowNonInteractableEntries)
 		{
 			if (mEntries[index]->isFullyInteractable())
 			{
-				mSelectedEntryIndex = index;
+				mSelectedEntryIndex = (int)index;
 				return true;
 			}
 		}
